requirements.txt
asgiref==3.8.1
Django==5.0.7
django-autoslug==1.9.9
et-xmlfile==1.1.0
openpyxl==3.1.5
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
pip install django-import-export
ohirida